Hey,
OK, this is perfectionism. I had http://jeroen.drop.org/ validated once again, and noticed an error with a </P>, it said that the paragraph wasn't opened. But it occured near a story as we do <P> check_output(story->abstract) </P> I was pretty certain the paragraph was open. :) OK, so it wasn't the theme's fault, neither was it the engine. Then I went to take a look at the actual story and what did I see. It used a BLOCKQUOTE. After deleting the BLOCKQUOTE, messing around with the HTML, trying to insert it again. It always gave me the error when it was inserted... What's my conclusion? BLOCKQUOTE isn't allowed inside a paragraph...
Solution? We could take the option to use the blockquote HTML tag in a story or comment out of drupal. But I really like blockquote. So I now changed <P> check_output(story->abstract) </P> to <BR> check_output(story->abstract) </BR> because, basicly, that does the same thing. I'd advise you to do so too, I might do it for you if you please. Not that I really care about it, but lately we've got some queer w3-lovers on our back. Let's shake them off, allright? ;)

Jeroen.